refactor(ui): reorg address book elements
