+++ apps/mobile/src/features/AddressBook/AddressBook.container.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { useAppSelector } from '@/src/store/hooks'
-import { AddressBookView } from '@/src/features/AddressBook/components/AddressBookView'
+import { AddressBookView } from '@/src/features/AddressBook/AddressBookView'
 import { selectAllContacts } from '@/src/store/addressBookSlice'
 
 export const AddressBookContainer = () => {
